requirements.txt
python-dotenv==1.1.0
requests==2.32.3
pandas-gbq==0.28.0